refactor: use kebab case for all file names
