+++ src/components/content-box/index.tsx
