+++ src/components/content-box/styles.tsx
