+++ src/components/icons/index.tsx
@@ -1,10 +1,10 @@
 import FacebookIcon from './facebook';
 import ForboleIcon from './forbole';
-import GithubIcon from './github';
-import LinkedinIcon from './linkedIn';
+import GithubIcon from './git-hub';
+import LinkedinIcon from './linked-in';
 import MediumIcon from './medium';
 import TelegramIcon from './telegram';
-import YoutubeIcon from './youtube';
+import YoutubeIcon from './you-tube';
 import TwitterIcon from './twitter';
 
 export {
