+++ src/components/index.ts
@@ -1,9 +1,9 @@
 import Layout from './layout';
 import Footer from './footer';
 import Nav from './nav';
-import SectionBox from './section_box';
-import SectionLimit from './section_limit';
-import ContentBox from './content_box';
+import SectionBox from './section-box';
+import SectionLimit from './section-limit';
+import ContentBox from './content-box';
 
 export {
   Layout,
