+++ src/components/section-box/styles.tsx
