+++ src/components/section-box/types.ts
