+++ src/screens/faq/components/menu-mobile/hooks.ts
