+++ src/screens/faq/components/menu-mobile/index.tsx
