+++ src/screens/faq/components/menu-mobile/styles.tsx
