+++ src/screens/privacy-policy/index.tsx
