+++ src/screens/privacy-policy/styles.tsx
