+++ src/screens/privacy-policy/utils.ts
