+++ src/screens/terms-and-conditions/index.tsx
