+++ src/screens/terms-and-conditions/styles.tsx
