+++ src/screens/terms-and-conditions/utils.ts
